+++ packages/haiku-formats/src/exporters/bodymovin/bodymovinExporter.ts
@@ -7,7 +7,8 @@ import * as Template from 'haiku-serialization/src/bll/Template';
 import * as difference from 'lodash/difference';
 import * as flatten from 'lodash/flatten';
 import * as mapKeys from 'lodash/mapKeys';
-import {Exporter} from '..';
+import {ExporterInterface} from '..';
+import BaseExporter from '../BaseExporter';
 
 import {SvgTag} from '../../svg/enums';
 import {
@@ -72,7 +73,7 @@ import {
 
 let bodymovinVersion: Maybe<string>;
 
-export class BodymovinExporter implements Exporter {
+export class BodymovinExporter extends BaseExporter implements ExporterInterface {
   /**
    * The out-point (last frame) for the animation.
    * @type {number}
@@ -922,31 +923,6 @@ export class BodymovinExporter implements Exporter {
   }
 
   /**
-   * Internal method for visiting every timeline and applying a callback to it.
-   */
-  private visitAllTimelines(callback: (timeline: any) => void) {
-    for (const timelineId in this.bytecode.timelines) {
-      for (const haikuId in this.bytecode.timelines[timelineId]) {
-        if (/^__/.test(haikuId)) {
-          continue;
-        }
-        callback(this.bytecode.timelines[timelineId][haikuId]);
-      }
-    }
-  }
-
-  /**
-   * Internal method for visiting every timeline property and applying a callback to it.
-   */
-  private visitAllTimelineProperties(callback: (timeline: any, property: string) => void) {
-    this.visitAllTimelines((timeline) => {
-      for (const property in timeline) {
-        callback(timeline, property);
-      }
-    });
-  }
-
-  /**
    * Normalizes keyframes so we can use 60fps without major modifications to direct parsing logic.
    *
    * Bodymovin uses keyframe-based transitions.
@@ -1218,7 +1194,8 @@ export class BodymovinExporter implements Exporter {
     return '{}';
   }
 
-  constructor(private bytecode) {
+  constructor(protected bytecode) {
+    super(bytecode);
     // If not already known, get the Bodymovin version.
     if (!bodymovinVersion) {
       bodymovinVersion = getBodymovinVersion();