+++ packages/haiku-sdk-creator/src/bll/Project.ts
@@ -1,12 +1,10 @@
-import { inkstone } from '@haiku/sdk-inkstone';
-import { client as sdkClient } from '@haiku/sdk-client';
-import { MaybeAsync } from '../envoy';
-
-import { Registry } from '../dal/Registry';
+import {client as sdkClient} from '@haiku/sdk-client';
+import {inkstone} from '@haiku/sdk-inkstone';
+import {MaybeAsync} from '../envoy';
 
 export interface Project {
-  setIsPublic: (uniqueId:string, isPublic:boolean) => MaybeAsync<inkstone.project.Project>;
-  getProjectDetail: (uniqueId:string) => Promise<inkstone.project.Project>;
+  setIsPublic: (uniqueId: string, isPublic: boolean) => MaybeAsync<boolean>;
+  getProjectDetail: (uniqueId: string) => Promise<inkstone.project.Project>;
 }
 
 export const PROJECT_CHANNEL = 'project';
@@ -21,7 +19,7 @@ export class ProjectHandler implements Project {
     }
   }
 
-  getProjectDetail(uniqueId: string) : Promise<inkstone.project.Project> {
+  getProjectDetail(uniqueId: string): Promise<inkstone.project.Project> {
     return new Promise<inkstone.project.Project>((resolve, reject) => {
       inkstone.project.getByUniqueId(sdkClient.config.getAuthToken(), uniqueId, (error, project) => {
         if (!error) {
@@ -33,22 +31,16 @@ export class ProjectHandler implements Project {
     });
   }
 
-  setIsPublic(uniqueId: string, isPublic: boolean) : Promise<inkstone.project.Project> {
-    return new Promise<inkstone.project.Project>((resolve, reject) => {
-      const params : inkstone.project.ProjectUpdateParams = {UniqueId: uniqueId};
-      if (isPublic) {
-        params.MakePublic = true;
-      } else {
-        params.MakePrivate = true;
-      }
-      inkstone.project.update(sdkClient.config.getAuthToken(), params, (error, project) => {
+  setIsPublic(uniqueId: string, isPublic: boolean): Promise<boolean> {
+    return new Promise<boolean>((resolve, reject) => {
+      const method = isPublic ? inkstone.project.makePublic : inkstone.project.makePrivate;
+      method(sdkClient.config.getAuthToken(), uniqueId, (error, response) => {
         if (!error) {
-          resolve(project);
+          resolve(response);
         } else {
           reject(error);
         }
       });
     });
   }
-
 }