+++ packages/haiku-serialization/test/bll/03_Asset.test.js
@@ -9,9 +9,8 @@ const PROJECT_MODEL_STUB = {
   }
 }
 
-test('Asset.assetsToDirectoryStructure', (t) => {
-  t.plan(10)
-  const assets = Asset.ingestAssets(PROJECT_MODEL_STUB, {
+const mockAssets = () =>{
+  return Asset.ingestAssets(PROJECT_MODEL_STUB, {
     'code/main/code.js': {
       relpath: 'code/main/code.js',
       abspath: path.join(__dirname, '..', 'projects', 'test-project', 'code/main/code.js'),
@@ -52,7 +51,24 @@ test('Asset.assetsToDirectoryStructure', (t) => {
       abspath: path.join(__dirname, '..', 'projects', 'test-project', 'designs/TEST.sketch.contents/slices/Slicey.svg'),
       dtModified: Date.now()
     },
+    'designs/ID-TEST.figma.contents/slices/Slicey.svg': {
+      relpath: 'designs/ID-TEST.figma.contents/slices/Slicey.svg',
+      abspath: path.join(__dirname, '..', 'projects', 'test-project', 'designs/ID-TEST.figma.contents/slices/Slicey.svg'),
+      dtModified: Date.now()
+    },
+    'designs/ID-TEST.figma.contents/groups/Slicey.svg': {
+      relpath: 'designs/ID-TEST.figma.contents/groups/Slicey.svg',
+      abspath: path.join(__dirname, '..', 'projects', 'test-project', 'designs/ID-TEST.figma.contents/groups/Slicey.svg'),
+      dtModified: Date.now()
+    },
   })
+}
+
+test('Asset.assetsToDirectoryStructure', (t) => {
+  t.plan(13)
+
+  const assets = mockAssets()
+
   t.ok(assets[0],'asset exists')
 
   const idx = (experimentIsEnabled(Experiment.MultiComponentControlsLibrary))
@@ -61,11 +77,66 @@ test('Asset.assetsToDirectoryStructure', (t) => {
 
   t.equal(assets[idx].kind, 'folder', 'base asset is folder')
   t.equal(assets[idx].type, 'container', 'base asset is container')
-  t.equal(assets[idx].children.length, 1, 'base asset has one child')
-  t.equal(assets[idx].children[0].kind, 'sketch', 'child asset is sketch')
+  t.equal(assets[idx].children.length, 2, 'base asset has two children')
+  t.equal(assets[idx].children[0].kind, 'sketch', 'first child asset is sketch')
+  t.equal(assets[idx].children[1].kind, 'figma', 'second child asset is figma')
   t.equal(assets[idx].children[0].type, 'container', 'child asset is container')
-  t.equal(assets[idx].children[0].children[0].relpath, 'designs/designs/TEST.sketch/slices', 'grandchild is slices folder')
+  t.equal(assets[idx].children[0].children[0].relpath, 'designs/designs/TEST.sketch/slices', 'sketch grandchild is slices folder')
+  t.equal(assets[idx].children[1].children[0].relpath, 'designs/designs/ID-TEST.figma/groups', 'figma grandchild is groups folder')
+  t.equal(assets[idx].children[1].children[1].relpath, 'designs/designs/ID-TEST.figma/slices', 'figma grandchild is slices folder')
   t.equal(assets[idx].children[0].children[0].kind, 'folder', 'grandchild is folder')
   t.equal(assets[idx].children[0].children[0].type, 'container', 'grandchild is container')
-  t.equal(assets[idx].dump(), "designs\n  designs/TEST.sketch\n    designs/designs/TEST.sketch/slices\n      designs/TEST.sketch.contents/slices/Dicey.svg\n      designs/TEST.sketch.contents/slices/Slicey.svg\n    designs/designs/TEST.sketch/artboards\n      designs/TEST.sketch.contents/artboards/Another Artboard.svg\n      designs/TEST.sketch.contents/artboards/Artboard.svg",'tree looks ok')
+  t.equal(assets[idx].dump(), "designs\n  designs/TEST.sketch\n    designs/designs/TEST.sketch/slices\n      designs/TEST.sketch.contents/slices/Dicey.svg\n      designs/TEST.sketch.contents/slices/Slicey.svg\n    designs/designs/TEST.sketch/artboards\n      designs/TEST.sketch.contents/artboards/Another Artboard.svg\n      designs/TEST.sketch.contents/artboards/Artboard.svg\n  designs/ID-TEST.figma\n    designs/designs/ID-TEST.figma/groups\n      designs/ID-TEST.figma.contents/groups/Slicey.svg\n    designs/designs/ID-TEST.figma/slices\n      designs/ID-TEST.figma.contents/slices/Slicey.svg",'tree looks ok')
+})
+
+test('Asset.assetsToDirectoryStructure detects sketch assets without exported SVG files', (t) => {
+  t.plan(2)
+
+  const assets = Asset.ingestAssets(PROJECT_MODEL_STUB, {
+    'designs/TEST.sketch': {
+      relpath: 'designs/TEST.sketch',
+      abspath: path.join(__dirname, '..', 'projects', 'test-project', 'designs/TEST.sketch'),
+      dtModified: Date.now()
+    },
+  })
+
+  t.equal(assets[0].children.length, 1, 'base asset has a child')
+  t.equal(assets[0].children[0].kind, 'sketch', 'child asset is sketch')
+})
+
+test('Asset.getAssetInfo', (t) => {
+  t.plan(4)
+
+  const assets = mockAssets()
+  const sketchAsset = assets[0].children[0].children[0].children[0]
+  const figmaAsset = assets[0].children[1].children[0].children[0]
+  const sketchAssetInfo = sketchAsset.getAssetInfo()
+  const figmaAssetInfo = figmaAsset.getAssetInfo()
+
+  t.equal(sketchAssetInfo.generator, 'sketch')
+  t.equal(sketchAssetInfo.generatorRelpath, 'designs/TEST.sketch')
+  t.equal(figmaAssetInfo.generator, 'figma')
+  t.equal(figmaAssetInfo.generatorRelpath, 'designs/ID-TEST.figma')
+})
+
+test('Asset.isSketchFile', (t) => {
+  t.plan(2)
+
+  const assets = mockAssets()
+  const sketchAsset = assets[0].children[0]
+  const figmaAsset = assets[0].children[1]
+
+  t.ok(sketchAsset.isSketchFile())
+  t.notOk(sketchAsset.isFigmaFile())
+})
+
+test('Asset.isFigmaFile', (t) => {
+  t.plan(2)
+
+  const assets = mockAssets()
+  const sketchAsset = assets[0].children[0]
+  const figmaAsset = assets[0].children[1]
+
+  t.ok(figmaAsset.isFigmaFile())
+  t.notOk(figmaAsset.isSketchFile())
 })