+++ packages/haiku-ui-common/src/react/ShareModal/index.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import Palette from '../../Palette';
 import {ModalWrapper, ModalHeader, ModalNotice} from '../Modal';
 import {RevealPanel} from '../RevealPanel';
 import {ProjectShareDetails} from './ProjectShareDetails';
@@ -34,6 +33,7 @@ export interface PropTypes {
   organizationName: string;
   projectUid: string;
   sha: string;
+  mixpanel: object;
 }
 
 export interface StateTypes {
@@ -59,7 +59,7 @@ export class ShareModal extends React.Component<PropTypes, StateTypes> {
 
     this.state = {
       showDetail: false,
-      isPublic: true, // default true so we don't ever accidentally tell the user their projects are more private than they are
+      isPublic: undefined,
       showTooltip: false,
       isPublicKnown: false,
     };
@@ -79,9 +79,8 @@ export class ShareModal extends React.Component<PropTypes, StateTypes> {
 
         // if IsPublic is undefined, it's never been published before.  toggle it true on first publish.
         if (proj.IsPublic === null || proj.IsPublic === undefined) {
-          (nextProps.envoyProject.setIsPublic(nextProps.projectUid, true) as Promise<inkstone.project.Project>).then((updatedProj) => {
-            this.setState({isPublic: updatedProj.IsPublic});
-          });
+          (nextProps.envoyProject.setIsPublic(nextProps.projectUid, true) as Promise<boolean>).then(() => {});
+          this.setState({isPublic: true});
         } else {
           this.setState({isPublic: proj.IsPublic, isPublicKnown: true});
         }
@@ -89,8 +88,13 @@ export class ShareModal extends React.Component<PropTypes, StateTypes> {
     }
   }
 
-  showDetails (selectedEntry: string) {
+  showDetails (selectedEntry) {
     this.setState({selectedEntry, showDetail: true});
+    this.props.mixpanel.haikuTrack('install-options', {
+      from: 'app',
+      event: 'show-single-option',
+      option: selectedEntry.entry,
+    });
   }
 
   hideDetails () {
@@ -103,10 +107,8 @@ export class ShareModal extends React.Component<PropTypes, StateTypes> {
       const desiredState = !this.state.isPublic;
       const project = props.envoyProject;
 
-      (project.setIsPublic(props.projectUid, desiredState) as Promise<inkstone.project.Project>).then((proj : inkstone.project.Project) => {
-        this.setState({isPublic: proj.IsPublic, isPublicKnown: true});
-      });
-
+      (project.setIsPublic(props.projectUid, desiredState) as Promise<boolean>).then(() => {});
+      this.setState({isPublic: desiredState, isPublicKnown: true});
     } else {
       // TODO:  trigger toast
       console.error('Could not set project privacy settings.  Please contact support@haiku.ai');
@@ -126,6 +128,7 @@ export class ShareModal extends React.Component<PropTypes, StateTypes> {
       organizationName,
       sha,
       projectUid,
+      mixpanel,
     } = this.props;
 
     return (
@@ -141,6 +144,7 @@ export class ShareModal extends React.Component<PropTypes, StateTypes> {
             isProjectInfoFetchInProgress={isProjectInfoFetchInProgress}
             isSnapshotSaveInProgress={isSnapshotSaveInProgress}
             isPublic={this.state.isPublic}
+            mixpanel={mixpanel}
             togglePublic={() => this.togglePublic()}
           />
         </ModalHeader>
@@ -152,6 +156,7 @@ export class ShareModal extends React.Component<PropTypes, StateTypes> {
               isSnapshotSaveInProgress={isSnapshotSaveInProgress}
               snapshotSyndicated={snapshotSyndicated}
               snapshotPublished={snapshotPublished}
+              mixpanel={mixpanel}
               onOptionClicked={(selectedEntry) => {
                 this.showDetails(selectedEntry);
               }}
@@ -165,6 +170,7 @@ export class ShareModal extends React.Component<PropTypes, StateTypes> {
               organizationName={organizationName}
               projectUid={projectUid}
               sha={sha}
+              mixpanel={mixpanel}
               onHide={() => {
                 this.hideDetails();
               }}